Split metrics collector
