+++ src/application/metrics/ddns.rs
@@ -1,6 +1,14 @@
-use prometheus_client::encoding::text::Encode;
+use prometheus_client::{
+    encoding::text::Encode,
+    registry::Registry,
+    metrics::{family::Family, gauge::Gauge},
+};
 
-use crate::domain::ddns::DdnsStatus;
+use crate::{
+    application::metrics::Collector,
+    domain::ddns::{DdnsStatus, DdnsUpdateStatus},
+    service::ddns::DdnsStatusResult,
+};
 
 #[derive(Clone, Hash, PartialEq, Eq, Encode)]
 pub struct DdnsStatusLabel {
@@ -18,6 +26,25 @@ impl From<DdnsStatus> for DdnsStatusLabel {
             interface_name,
             ip_address,
             hostname,
+        }
+    }
+}
+
+impl Collector for DdnsStatusResult {
+    fn collect(self, registry: &mut Registry) {
+        let ddns_status = Family::<DdnsStatusLabel, Gauge>::default();
+        registry.register(
+            "edgerouter_dynamic_dns_status",
+            "Result of DDNS update",
+            Box::new(ddns_status.clone()),
+        );
+
+        for status in self {
+            let value = match status.update_status {
+                Some(DdnsUpdateStatus::Good | DdnsUpdateStatus::NoChange | DdnsUpdateStatus::NoError) => 1,
+                _ => 0,
+            };
+            ddns_status.get_or_create(&status.into()).set(value);
         }
     }
 }
