+++ src/application/metrics/version.rs
@@ -1,6 +1,14 @@
-use prometheus_client::encoding::text::Encode;
+use prometheus_client::{
+    encoding::text::Encode,
+    registry::Registry,
+    metrics::{family::Family, gauge::Gauge},
+};
 
-use crate::domain::version::Version;
+use crate::{
+    application::metrics::Collector,
+    domain::version::Version,
+    service::version::VersionResult,
+};
 
 #[derive(Clone, Hash, PartialEq, Eq, Encode)]
 pub struct VersionLabel {
@@ -16,5 +24,18 @@ impl From<Version> for VersionLabel {
             build_id: v.build_id,
             model: v.hw_model,
         }
+    }
+}
+
+impl Collector for VersionResult {
+    fn collect(self, registry: &mut Registry) {
+        let info = Family::<VersionLabel, Gauge>::default();
+        registry.register(
+            "edgerouter_info",
+            "Version info",
+            Box::new(info.clone()),
+        );
+
+        info.get_or_create(&self.into()).set(1);
     }
 }
