+++ src/di/container.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use crate::{
-    application::server::Engine,
+    application::{metrics::MetricsController, server::Engine},
     infrastructure::{
         cmd::{
             parser::{
@@ -29,14 +29,16 @@ impl Application {
     pub async fn start(config: &Config) -> anyhow::Result<()> {
         let engine = Engine::new(
             config.port,
-            None,
-            None,
-            BGPRunner::new(&config.vtysh_command, BGPParser),
-            DdnsRunner::new(&config.op_ddns_command, DdnsParser),
-            LoadBalanceRunner::new(&config.op_command, LoadBalanceParser),
-            PPPoERunner::new(&config.op_command, PPPoEParser),
-            VersionRunner::new(&config.op_command, VersionParser),
-        )?;
+            config.tls_cert.clone(),
+            config.tls_key.clone(),
+            MetricsController::new(
+                BGPRunner::new(&config.vtysh_command, BGPParser),
+                DdnsRunner::new(&config.op_ddns_command, DdnsParser),
+                LoadBalanceRunner::new(&config.op_command, LoadBalanceParser),
+                PPPoERunner::new(&config.op_command, PPPoEParser),
+                VersionRunner::new(&config.op_command, VersionParser),
+            ),
+        );
         Ok(Arc::new(engine).start().await)
     }
 }
